saperated api calls from the reducer in order to test the async actions
